refactor: split CloudBeesAuthPage into subcomponents
Break the 1012-line monolith into focused presentational components:
- ModeSelector (step 1): connection mode selection
- CredentialForms (step 2): OC, single controller, and PAT forms
- WebhookSetup (step 3): webhook URL display and Jenkinsfile snippet
- ConnectedDashboard: stats, job health, RCA toggle, deployments, controllers

Parent retains all state, effects, and handler logic.
